Updated post like commands queries

+++ src/Services/Posts/InstaConnect.Posts.Infrastructure/Features/PostLikes/Helpers/PostLikeService.cs
@@ -1,4 +1,7 @@
-﻿using InstaConnect.Common.Domain.Models.Pagination;
+﻿using System.Threading;
+
+using InstaConnect.Common.Domain.Models.Pagination;
+using InstaConnect.Common.Exceptions.Users;
 using InstaConnect.Posts.Domain.Features.PostLikes.Exceptions;
 using InstaConnect.Posts.Domain.Features.PostLikes.Models.Filters;
 
@@ -40,19 +43,33 @@ internal class PostLikeService : IPostLikeService
         return postLike;
     }
 
-    public void Add(Post post, string userId)
+    public async Task<PostLike> AddAsync(Post post, string userId, CancellationToken cancellationToken)
     {
+        var existingPostLike = await _postLikeWriteRepository.GetByUserIdAndPostIdAsync(userId, post.Id, cancellationToken);
+
+        if (existingPostLike != null)
+        {
+            throw new PostLikeAlreadyExistsException();
+        }
+
         var postLike = _postLikeFactory.Get(post.Id, userId);
         _postLikeWriteRepository.Add(postLike);
+
+        return postLike;
     }
 
-    public async Task DeleteAsync(Post post, string id, CancellationToken cancellationToken)
+    public async Task DeleteAsync(Post post, string id, string userId, CancellationToken cancellationToken)
     {
         var postLike = await _postLikeReadRepository.GetByIdAsync(id, cancellationToken);
 
         if (postLike == null)
         {
             throw new PostLikeNotFoundException();
+        }
+
+        if (userId != postLike.UserId)
+        {
+            throw new UserForbiddenException();
         }
 
         _postLikeWriteRepository.Delete(postLike);
